# [Java.Interop] remove IEnumerable iterators called at startup