@jonathanpeppers 2020-01-08
Probably should wait on the results of the PR build here: https://github.com/xamarin/xamarin-android/pull/4102