@jonathanpeppers 2020-01-09
This change on the xamarin-android side seems to be OK: https://build.azdo.io/3368801