@jonathanpeppers 2020-01-13
An updated xamarin-android build is running here: https://build.azdo.io/3380356