@jonpryor 2020-01-13
Looks like we'll need to rebase the xamarin/xamarin-android#4102 PR and rebase atop xamarin/xamarin-android@e22be491, as it's getting lots of build failures due to `NuGet.Common`.