@jonpryor 2020-01-14
My concern with this is that it's a semantic API break:

*Currently* a subclass would override `Java.Interop.JniRuntime.JniTypeManager.GetSimpleReferences()`, and a call to `JniTypeManager.GetTypeSignature()` will call `GetSimpleReferences()`.

The [`xamarin-android` repo currently does this](https://github.com/xamarin/xamarin-android/blob/e22be4912e498c705e89acc4e2fc3af8a0a9a8c4/src/Mono.Android/Android.Runtime/AndroidRuntime.cs#L232).

After this PR, `GetTypeSignature()` will no longer call `GetSimpleReferences()`.  This *would* break xamarin-android, except that xamarin/xamarin-android#4102 adds a `GetSimpleReferences()` override as well.

Now on the "plus" side, (in theory) nobody is actually using this API, so changing semantics shouldn't actually break anybody.

That said, it instead results in documentation confusion.  We'll have this class:

```csharp
partial class JniTypeManager {
    public JniTypeSignature GetTypeSignature (Type type);
    protected virtual string GetSimpleReference (Type type);

    public IEnumerable<JniTypeSignature> GetTypeSignatures (Type type);
    protected virtual IEnumerable<string> GetSimpleReferences (Type type);
}
```

Which method needs to be overridden?  One?  Both?

I think it would be better if `GetSimpleReference()` *delegated through* `GetSimpleReferences()`:

```csharp
protected virtual string GetSimpleReference (Type type)
{
    return GetSimpleReferences (type).FirstOrDefault ();
}
```

That way existing code will not require a *semantic* change, but only a *performance* change.