@jonathanpeppers 2020-01-14
An updated xamarin-android build is running here: https://build.azdo.io/3384618